Create use-delete-site to re-use code between settings and context menu
