+++ src/components/delete-site.tsx
@@ -1,11 +1,8 @@
-import * as Sentry from '@sentry/electron/renderer';
 import { MenuItem } from '@wordpress/components';
-import { __, sprintf } from '@wordpress/i18n';
+import { __ } from '@wordpress/i18n';
 import { useI18n } from '@wordpress/react-i18n';
+import { useDeleteSite } from 'src/hooks/use-delete-site';
 import { useSiteDetails } from 'src/hooks/use-site-details';
-import { getIpcApi } from 'src/lib/get-ipc-api';
-
-const MAX_LENGTH_SITE_TITLE = 35;
 
 type DeleteSiteProps = {
 	onClose: () => void;
@@ -13,51 +10,8 @@ type DeleteSiteProps = {
 
 const DeleteSite = ( { onClose }: DeleteSiteProps ) => {
 	const { __ } = useI18n();
-	const { selectedSite, deleteSite, isDeleting } = useSiteDetails();
-
-	const handleDeleteSite = async () => {
-		if ( ! selectedSite ) {
-			return;
-		}
-
-		const DELETE_BUTTON_INDEX = 0;
-		const CANCEL_BUTTON_INDEX = 1;
-
-		const trimmedSiteTitle = getTrimmedSiteTitle( selectedSite.name );
-
-		const { response, checkboxChecked } = await getIpcApi().showMessageBox( {
-			type: 'warning',
-			message: sprintf( __( 'Delete %s' ), trimmedSiteTitle ),
-			detail: __(
-				'The site’s database will be lost. Including all posts, pages, comments, and media.'
-			),
-			buttons: [ __( 'Delete site' ), __( 'Cancel' ) ],
-			cancelId: CANCEL_BUTTON_INDEX,
-			checkboxLabel: __( 'Delete site files from my computer' ),
-			checkboxChecked: true,
-		} );
-
-		if ( response === DELETE_BUTTON_INDEX ) {
-			try {
-				await deleteSite( selectedSite.id, checkboxChecked );
-			} catch ( error ) {
-				getIpcApi().showErrorMessageBox( {
-					title: __( 'Deletion failed' ),
-					message: sprintf(
-						__( "We couldn't delete the site '%s'. Please try again" ),
-						trimmedSiteTitle
-					),
-					error,
-				} );
-				Sentry.captureException( error );
-			}
-		}
-	};
-
-	const getTrimmedSiteTitle = ( name: string ) =>
-		name.length > MAX_LENGTH_SITE_TITLE
-			? `${ name.substring( 0, MAX_LENGTH_SITE_TITLE - 3 ) }…`
-			: name;
+	const { selectedSite, isDeleting } = useSiteDetails();
+	const { handleDeleteSite } = useDeleteSite();
 
 	const isSiteDeletionDisabled = ! selectedSite || isDeleting;
 
@@ -65,11 +19,11 @@ const DeleteSite = ( { onClose }: DeleteSiteProps ) => {
 		<MenuItem
 			aria-disabled={ isSiteDeletionDisabled }
 			onClick={ () => {
-				if ( isSiteDeletionDisabled ) {
+				if ( isSiteDeletionDisabled || ! selectedSite ) {
 					return;
 				}
 				onClose();
-				void handleDeleteSite();
+				void handleDeleteSite( selectedSite.id, selectedSite.name );
 			} }
 			isDestructive
 			disabled={ isSiteDeletionDisabled }
